fix: Out of range value for Phone Number

+++ Backend/Database/db_scripts/up.sql
@@ -4,7 +4,7 @@ CREATE TABLE user (
     password varchar(100) NOT NULL ,
     address varchar(100) NOT NULL ,
     birthday date NOT NULL ,
-    phone INTEGER(10),
+    phone NUMERIC(10),
     role enum('commuter','admin') DEFAULT 'commuter' NOT NULL ,
     PRIMARY KEY (email),
     CONSTRAINT validEmail CHECK (email LIKE '%@%.%'),
